% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation.R
\name{augment}
\alias{augment}
\title{Augmentation of simple_brix model}
\usage{
augment(object)
}
\arguments{
\item{object}{object of class "brix_simple", see \code{\link{brix_simple}}.}
}
\value{
Returns an object of \code{\link{class}} "augment". Using the print command displays test results from both
the Shaprio-Wilks Normality test amd the Kolmogorov-Smirnov one-sample test for the standard normal
transformed augmented data. See \code{\link{shapiro.test}} and \code{\link{ks.test}}. Using the plot
command displays standard normal Q-Q plots for the standard normal transformed augmented data by default.
}
\description{
Augments simple_brix model with unobserved claims (losses smaller than the deductibles)
for claims model distribution testing.
}
\details{
Using the Poisson distribution for augmentation of unobserved claims. (Might implement a Geometric method).
}
\examples{
b <- brix_simple(optimizer = optim,
                 fn = normal_poisson_log_likelihood,
                 par = c(1, 1, 1),
                 control = list(fnscale = -1, maxit = 1e6),
                 hessian = TRUE,
                 X = normal_poisson_x, N = normal_poisson_n)

a <- augment(object = b)
print(a)
plot(a)
plot(a, augment = FALSE)

}
