% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brix.R
\name{brix}
\alias{brix}
\title{Fitting regression models using maximum likelihood}
\usage{
brix(start_value, ll, formula = NULL, response = NULL, design = NULL,
  data, ...)
}
\arguments{
\item{start_value}{real valued vector of start values for the parameters.}

\item{ll}{log-likelihood function, see \code{\link{log_likelihood}}.}

\item{formula}{a symbolic description of the of the model to be fitted, see 'Details' for more information.}

\item{response}{name of response variable}

\item{design}{design/model matrix, see e.g. \code{\link{model.matrix}}}

\item{data}{the dataset}

\item{...}{additional arguments passed to optim}
}
\description{
Estimates parameters by the method of maximum likelihood in the chosen model along with other useful quantities.
}
\details{
The maximization of the log-likelihood function specified is done using the \code{\link{optim}}.
}
\examples{
brix(start_value = c(0, 0),
     ll = log_likelihood(fam = poisson(link = 'log')),
     formula = claim ~ age,
     data = policy_data)

brix(start_value = c(1, 1),
     ll = log_likelihood(fam = poisson(link = 'identity')),
     formula = claim ~ age,
     data = policy_data)

brix(start_value = c(0, 0, 0.1),
     ll = log_likelihood(fam = zmpoisson(link = 'log')),
     formula = claim ~ age,
     data = policy_data)

brix(start_value = c(1, 1, 0.1),
     ll = log_likelihood(fam = zmpoisson(link = 'identity')),
     formula = claim ~ age,
     data = policy_data)

brix(start_value = c(0, 0, 0.1),
     ll = log_likelihood(fam = zmpoisson(link = 'log')),
     response = 'claim',
     design = model.matrix(claim ~ age, data = policy_data),
     data = policy_data)
}
