% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_poisson_log_likelihood.R
\name{brix_simple}
\alias{brix_simple}
\title{Simple optimizer}
\usage{
brix_simple(optimizer, fn, ...)
}
\arguments{
\item{optimizer}{recommended choice is \code{optim}.}

\item{fn}{function to be optimized.}

\item{...}{additional arguments passed to the \code{optimizer}, see \code{examples}.}
}
\value{
Returns an object of \code{\link{class}} "brix_simple". The function \code{summary} can
also be used to print a summary of the model.

An object of class "brix" is a list containing at the function call and the results from
the optimizer. The optimal parameters are named if applicable. The summary contains at
least the following components:

\item{call}{the function call.}
\item{coefficients}{the optimal coefficents.}
\item{value}{the optimal value of the log-likelihood function.}
\item{AIC}{the AIC of the log-likelihood.}
}
\description{
A simple optimizer for log-likelihoods or other functions.
}
\details{
The optimizer must always be specified first.
The function to be optimized must always be specified second.
See e.g. \code{\link{optim}} for the optimizer.
}
\examples{
# Using optim for optimization
model <- brix_simple(optimizer = optim,
                     # function to be optimized
                     fn = normal_poisson_log_likelihood,
                     # arguments passed to fn
                     X = normal_poisson_x, N = normal_poisson_n,
                     # additional arguments passed to optim
                     par = rep(1, 3),
                     control = list(fnscale = -1),
                     method = "Nelder-Mead",
                     hessian = TRUE)
model
summary(model)

}
