% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{claim_data}
\alias{claim_data}
\title{Claim data}
\format{A data frame with 424 observations (rows) and 9 variables (columns):
\describe{
  \item{claim}{Total loss amount, including the deductible in USD,
  after taking the logarithm.}
  \item{deductible}{The deductible in USD, after taking the logarithm.}
  \item{age}{Age of the ship in years.}
  \item{brt}{Gross tonnage of the ship, after taking the logarithm.}
  \item{dwt}{Deadweight tonnage of the ship, after taking the logarithm.}
  \item{value}{Sum insured in USD, after taking the logrithm.}
  \item{hp}{Horsepowers.}
  \item{year}{Year of insurance.}
  \item{code}{Type of ship, 1 is bulk and 2 is cargo.}
}}
\usage{
claim_data
}
\description{
Claim data from insurance company.
}
\keyword{datasets}
