% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Generic log-likelihood for regression purposes.}
\usage{
log_likelihood(fam)
}
\arguments{
\item{fam}{a description of the the distribution and link function
to specify the log-likelihood, see "Details".}
}
\description{
Generic log-likelihood for regression purposes to be used with \code{\link{brix}}.
}
\details{
The options for \code{fam} argument are either \code{poisson} or
\code{zmpoisson}.
}
\examples{
x <- policy_data$claim

do.call(what = log_likelihood(fam = poisson(link = 'log'))$fun,
        args = list(beta = rep(1, length(x)), z = diag(x), x = x))

do.call(what = log_likelihood(fam = poisson(link = 'log'))$fun,
        args = list(beta = rep(1, length(x)), z = diag(x), x = x))

model1 <- brix(start_value = c(0, 0),
               ll = log_likelihood(fam = poisson(link = 'log')),
               formula = claim ~ age,
               data = policy_data)
model1
summary(model1)

model2 <- brix(start_value = c(0, 0, 0, 0, 0, 0.1),
               ll = log_likelihood(fam = zmpoisson(link = 'log')),
               formula = claim ~ age + brt + dwt + hp,
               data = policy_data)
model2
summary(model2)
}
\seealso{
\code{\link{log_likelihood_poisson}}, \code{\link{log_likelihood_zmpoisson}}.
}
