% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{log_likelihood_poisson}
\alias{log_likelihood_poisson}
\alias{log_likelihood_poisson}
\alias{log_likelihood_poisson_log}
\alias{log_likelihood_poisson_id}
\title{Poisson log-likelihood for regression purposes.}
\usage{
log_likelihood_poisson(beta, z, x, link = "log", sum = TRUE)

log_likelihood_poisson_log(beta, z, x, sum = TRUE)

log_likelihood_poisson_id(beta, z, x, sum = TRUE)
}
\arguments{
\item{beta}{\code{numeric} vector of parameters.}

\item{z}{\code{matrix}, model/design matrix, see e.g. \code{\link{model.matrix}}.}

\item{x}{\code{numeric} vector of (non-negative integer) quantiles.}

\item{link}{link function, the default is the logarithm,
but the identity is also avalible.}

\item{sum}{\code{logical}, default is \code{TRUE}.}
}
\value{
The output is a \code{numeric} value. Make sure the dimension of
\code{beta} and \code{z} are legal for matrix multiplication, see the examples.
}
\description{
Poisson log-likelihood for regression purposes to be used with
\code{\link{brix}}, see also \code{\link{log_likelihood}}.
}
\examples{
x <- policy_data$claim

log_likelihood_poisson(beta = c(0, 0),
                       z = model.matrix(claim ~ age, data = policy_data),
                       x = x)

log_likelihood_poisson(beta = rep(1, length(x)),
                       z = diag(x),
                       x = x)

do.call(what = log_likelihood(fam = poisson(link = "log"))$fun,
        args = list(beta = rep(1, length(x)), z = diag(x), x = x))
}
