% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{log_likelihood_weibull}
\alias{log_likelihood_weibull}
\alias{log_likelihood_weibull}
\alias{log_likelihood_weibull_log}
\alias{log_likelihood_weibull_id}
\title{Weibull log-likelihood for regression purposes}
\usage{
log_likelihood_weibull(beta, z, x, d = 0, link = "log", sum = TRUE)

log_likelihood_weibull_log(beta, z, x, d = 0, sum = TRUE)

log_likelihood_weibull_id(beta, z, x, d = 0, sum = TRUE)
}
\arguments{
\item{beta}{\code{numeric} vector of parameters.}

\item{z}{\code{matrix}, model/design matrix, see e.g. \code{\link{model.matrix}}.}

\item{x}{\code{numeric} vector of (non-negative integer) quantiles.}

\item{d}{\code{numeric} vector of deductibles.}

\item{link}{link function, the default is the logarithm, but the identity
is also avalible.}

\item{sum}{\code{logical}, default is \code{TRUE}.}
}
\value{
The output is a \code{numeric} value, make sure the dimension of
\code{beta} and \code{z} are legal for matrix multiplication, see the examples.
}
\description{
Weibull log-likelihood for regression purposes to be used
with \code{\link{brix}}, see also \code{\link{log_likelihood}}.
}
\examples{
x <- claim_data$claim

log_likelihood_weibull(beta = c(0.5, 0.5, 2), # 2 is the gamma parameter
                         z = model.matrix(claim ~ age, data = claim_data),
                         x = x)

log_likelihood_weibull(beta = c(rep(0, length(x)), 1),
                         z = diag(x),
                         x = x)

do.call(what = log_likelihood(fam = weibull(link = "log"))$fun,
        args = list(beta = c(rep(0, length(x)), 1), z = diag(x), x = x))
}
