% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmpoisson.R
\name{log_likelihood_zmpoisson}
\alias{log_likelihood_zmpoisson}
\alias{log_likelihood_zmpoisson}
\alias{log_likelihood_zmpoisson_log}
\alias{log_likelihood_zmpoisson_id}
\title{Zero-modified Poisson log-likelihood for regression purposes}
\usage{
log_likelihood_zmpoisson(beta, z, x, d = 0, link = "log", sum = TRUE)

log_likelihood_zmpoisson_log(beta, z, x, d = 0, sum = TRUE)

log_likelihood_zmpoisson_id(beta, z, x, d = 0, sum = TRUE)
}
\arguments{
\item{beta}{\code{numeric} vector of parameters.}

\item{z}{\code{matrix}, model/design matrix, see e.g. \code{\link{model.matrix}}.}

\item{x}{\code{numeric} vector of (non-negative integer) quantiles.}

\item{d}{\code{numeric} vector of deductibles.}

\item{link}{link function, the default is the logarithm, but the identity
is also avalible.}

\item{sum}{\code{logical}, default is \code{TRUE}.}
}
\value{
The output is a \code{numeric} value, make sure the dimension of
\code{beta} and \code{z} are legal for matrix multiplication, see the examples.
}
\description{
Zero-modified Poisson log-likelihood for regression purposes to be used
with \code{\link{brix}}, see also \code{\link{log_likelihood}}.
}
\examples{
x <- policy_data$claim

log_likelihood_zmpoisson(beta = c(0, 0, 0.5), # 0.5 is the p0 parameter
                         z = model.matrix(claim ~ age, data = policy_data),
                         x = x)

log_likelihood_zmpoisson(beta = c(rep(1, length(x)), 0.5),
                         z = diag(x),
                         x = x)

do.call(what = log_likelihood(fam = zmpoisson(link = "log"))$fun,
        args = list(beta = c(rep(1, length(x)), 0.5), z = diag(x), x = x))
}
