% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_poisson_log_likelihood.R
\name{normal_poisson_log_likelihood}
\alias{normal_poisson_log_likelihood}
\title{Normal-Poisson log-likelihood}
\usage{
normal_poisson_log_likelihood(parameter = NULL, X = NULL, N = NULL,
  check = 1)
}
\arguments{
\item{parameter}{\code{numeric}, vector of parameters of length 3 consisting of the mean and standard
deviation of the Normal component and the mean of the Poisson component. Default is \code{NULL}.}

\item{X}{\code{data.frame}, must have two columns named \code{claim} and \code{deductible}
for the Normal component of the log-likelihood. Default is \code{NULL}.}

\item{N}{\code{data.frame}, must have two columns \code{occurrence} and \code{deductible}
for the Poisson component of the log-likelihood. Default is \code{NULL}.}

\item{check}{default value is 1 and returns the log-likelihood, else returns the parameter
names of the model: \code{mean}, \code{sd} and \code{lambda}.}
}
\value{
Returns a \code{numeric} value or \code{character} vector of parameter names depending on
whether \code{check} = 1 or not.
}
\description{
Implementation of Normal-Poisson log-likelihood, taking deductibles into account.
}
\details{
Note that the row number of \code{X} and \code{N} need not be the same.
}
\examples{
normal_poisson_log_likelihood(c(1,2,3), X = normal_poisson_x, N = normal_poisson_n, check = 1)
normal_poisson_log_likelihood(c(1,2,3), X = normal_poisson_x, N = normal_poisson_n, check = 0)
normal_poisson_log_likelihood(check = 0)

}
