% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weibull}
\alias{weibull}
\title{Weibull family}
\usage{
weibull(link = "log")
}
\arguments{
\item{link}{link function, the default is the logarithm,
but the identity is also avalible.}
}
\description{
Family for use with \code{\link{brix}}, implementing regression for
Weibull data.
}
\examples{
weibull(link = "log")
weibull(link = "identity")
weibull()

}
