% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmpoisson.R
\name{zmpoisson}
\alias{zmpoisson}
\title{Zero-modified Poisson family}
\usage{
zmpoisson(link = "log")
}
\arguments{
\item{link}{link function, the default is the logarithm,
but the identity is also avalible.}
}
\description{
Family for use with \code{\link{brix}}, implementing regression for
zero-modified Poisson data.
}
\examples{
zmpoisson(link = "log")
zmpoisson(link = "identity")
zmpoisson()

}
