% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brix.R
\name{brix}
\alias{brix}
\title{Fitting regression models using maximum likelihood}
\usage{
brix(start_value, ll, formula = NULL, response = NULL, design = NULL,
  data, d = 0, ...)
}
\arguments{
\item{start_value}{a \code{numeric} vector of starting values for the parameters.}

\item{ll}{log-likelihood function, see \code{\link{log_likelihood}}.}

\item{formula}{\code{formula}, a symbolic description of the of the model to be fitted,
see e.g. \code{\link{formula}}.}

\item{response}{\code{character}, name of response variable.}

\item{design}{\code{matrix}, design/model matrix, see e.g. \code{\link{model.matrix}}.}

\item{data}{\code{data.frame}, the dataset used for regression.}

\item{d}{\code{numeric} vector of deductibles.}

\item{...}{additional arguments passed to optim.}
}
\value{
\code{brix} returns an object of \code{\link{class}} "brix". The function summary
can be used to print a summary of the model.

An object of class "brix" is a list containing at least the following components:

\item{par}{a named vector of maximum likelihood estimates.}
\item{value}{maximum log-likelihood value.}
\item{hessian}{estimate of the Hessian at the maximum log-likelihood value
 from \code{optim}.}
\item{z}{design/model matrix.}
\item{data}{data frame used for regression.}
\item{coef}{the same as \code{par}.}
\item{null.deviance}{the deviance for the null model, comparable with deviance.}
\item{deviance}{the deviance for the current model.}
\item{AIC, BIC}{The Akaike Information Criterion and the Bayesian Information
Criterion.}
}
\description{
Estimates parameters by the method of maximum likelihood in the chosen model
along with other useful quantities.
}
\details{
The maximization of the log-likelihood function specified is done using the
\code{\link{optim}} function. The \code{response} and \code{design} arguments
are meant to be used together instead of the \code{formula} argument, although using
\code{formula} is usually prefered.
}
\examples{
brix(start_value = c(0, 0),
     ll = log_likelihood(fam = poisson(link = 'log')),
     formula = claim ~ age,
     data = policy_data)

brix(start_value = c(1, 1),
     ll = log_likelihood(fam = poisson(link = 'identity')),
     formula = claim ~ age,
     data = policy_data)

brix(start_value = c(0, 0, 0.1),
     ll = log_likelihood(fam = zmpoisson(link = 'log')),
     formula = claim ~ age,
     data = policy_data)

brix(start_value = c(1, 1, 0.1),
     ll = log_likelihood(fam = zmpoisson(link = 'identity')),
     formula = claim ~ age,
     data = policy_data)

brix(start_value = c(0, 0, 0.1),
     ll = log_likelihood(fam = zmpoisson(link = 'log')),
     response = 'claim',
     design = model.matrix(claim ~ age, data = policy_data),
     data = policy_data)

brix(start_value = c(1, 1, 0.1),
     ll = log_likelihood(fam = weibull(link = "log")),
     formula = claim ~ age,
     data = claim_data)

 brix(start_value = c(1, 1, 0.1),
     ll = log_likelihood(fam = weibull(link = "identity")),
     formula = claim ~ age,
     data = claim_data)

 brix(start_value = c(1, 1, 0.1),
 ll = log_likelihood(fam = weibull(link = "log")),
 formula = claim ~ age,
data = claim_data,
d = claim_data$deductible)

}
